removed outdated classes form 52n-wps-ags module; restructured internal classes
