+++ 52n-wps-ags/src/main/java/org/n52/wps/ags/ArcToolboxProcessDelegator.java
@@ -43,10 +43,10 @@ import net.opengis.wps.x100.ProcessDescriptionType;
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.n52.wps.ags.feed.movingcode.AlgorithmParameterType;
-import org.n52.wps.ags.feed.movingcode.CommandLineParameter;
-import org.n52.wps.ags.feed.movingcode.MovingCodeObject;
-import org.n52.wps.ags.feed.movingcode.MovingCodeUtils;
+import org.n52.wps.ags.algorithmpackage.AlgorithmPackage;
+import org.n52.wps.ags.algorithmpackage.AlgorithmParameterType;
+import org.n52.wps.ags.algorithmpackage.CommandLineParameter;
+import org.n52.wps.ags.algorithmpackage.PackageUtils;
 import org.n52.wps.ags.workspace.AGSWorkspace;
 import org.n52.wps.io.data.GenericFileDataWithGT;
 import org.n52.wps.io.data.GenericFileDataConstants;
@@ -66,13 +66,13 @@ public class ArcToolboxProcessDelegator implements IAlgorithm{
 	private final File toolWorkspace;
 	private final File workspaceBase;
 
-	private MovingCodeObject mco;
+	private AlgorithmPackage mco;
 	protected List<String> errors;
 
 	private String[] toolParameters;
 
 
-	protected ArcToolboxProcessDelegator(MovingCodeObject templateMCO, File workspaceBase) throws IOException{
+	protected ArcToolboxProcessDelegator(AlgorithmPackage templateMCO, File workspaceBase) throws IOException{
 		this.workspaceBase = workspaceBase;
 		errors = new ArrayList<String>();
 		mco = templateMCO.createChild(workspaceBase);
@@ -117,7 +117,7 @@ public class ArcToolboxProcessDelegator implements IAlgorithm{
 					while (it.hasNext()){
 						IData currentItem = it.next();
 						// load as file and add to CommanLineParameter
-						cmdParam.addValue((MovingCodeUtils.loadSingleDataItem(currentItem, toolWorkspace)));
+						cmdParam.addValue((PackageUtils.loadSingleDataItem(currentItem, toolWorkspace)));
 					}
 				}
 			} else if (!wpsOutputID.equalsIgnoreCase("")){ // output only parameters !!ONLY SINGLE OUTPUT ITEMS SUPPORTED BY WPS!!
@@ -201,11 +201,11 @@ public class ArcToolboxProcessDelegator implements IAlgorithm{
 	}
 
 	public Class<?> getInputDataType(String id) {
-		return MovingCodeUtils.getInputDataType(mco, id);
+		return PackageUtils.getInputDataType(mco, id);
 	}
 
 	public Class<?> getOutputDataType(String id) {
-		return MovingCodeUtils.getOutputDataType(mco, id);
+		return PackageUtils.getOutputDataType(mco, id);
 	}
 
 	//delete the current workspace
