+++ 52n-wps-ags/src/main/java/org/n52/wps/ags/algorithmpackage/CommandLineParameter.java
@@ -26,7 +26,7 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
  * Public License for more details.
  */
-package org.n52.wps.ags.feed.movingcode;
+package org.n52.wps.ags.algorithmpackage;
 
 import java.util.LinkedList;
